Update missileGuidance.pde
i think the problem might have to do with the precision, t1 and t2 probably arent exactly equal, so i added a code to check if they are in the range of .01 and it seems to work. But i noticed that no matter what time we stop it at, our t1 and t2 values are equal to each other, so its always saying there is a hit when it shouldn't be.